requirements.txt
pandas
numpy
alpaca_py
datetime
asyncio
time
winsound
apscheduler
tables

